#135: separate edit content
