+++ lib/state/notes/notes_state.dart
@@ -21,7 +21,7 @@ final class NotesState extends Equatable {
   @override
   List<Object?> get props => [status, notes, error];
 
-  NoteModel getNote(String id) {
-    return notes.firstWhere((element) => element.id == id);
+  NoteModel? getNote(String? id) {
+    return notes.where((element) => element.id == id).firstOrNull;
   }
 }
