+++ lib/ui/pages/home/home_loading.dart
@@ -12,6 +12,7 @@ class HomeLoading extends StatelessWidget {
           crossAxisCount: 2,
           mainAxisSpacing: 12,
           crossAxisSpacing: 12,
+          childAspectRatio: NotedWidgetConfig.tileAspectRatio,
           children: List.filled(8, LoadingBox()),
         ),
       ),
